Fix crash when running with a wheel that doesn't support all the effects
Return DIERR_UNSUPPORTED when attempting to play force feedback effects
that do not exist on the hardware that is plugged in. Prevents crashing
the Unity editor when playing inside the editor, when there is a device
plugged in that does not support the springForce, damperForce, or
constantForce functions.

Also return false from the bool functions for stopping said effects.

Also make sure that all fields in the device struct are initialized to
zero when the device struct is created.