remove DataPoint collection
The DataPoint collection is not used in this project, and the collection
of large quantities of potentially garbage collectable data that never
gets released (or used), causes extreme spikes in garbage collection after
running the simulator for periods of time longer than a few minutes.

After removing the DataPoint collection, although garbage collection
is still undesireably frequent (string.Format every frame causes collection
to occur very frequently, in the version of .Net used in Unity currently),
the length of time that it takes for each collection is reduced from
spikes of hundreds of milliseconds, to only around 5-10ms per collection.